+++ app/src/main/kotlin/me/xizzhu/android/joshua/core/repository/BibleReadingRepository.kt
@@ -95,10 +95,6 @@ class BibleReadingRepository(private val localReadingStorage: LocalReadingStorag
     suspend fun readVerse(translationShortName: String, verseIndex: VerseIndex): Verse =
             localReadingStorage.readVerse(translationShortName, verseIndex)
 
-    suspend fun readVerse(translationShortName: String, parallelTranslations: List<String>,
-                          verseIndex: VerseIndex): Verse =
-            localReadingStorage.readVerse(translationShortName, parallelTranslations, verseIndex)
-
     suspend fun search(translationShortName: String, query: String): List<Verse> {
         Analytics.track(Analytics.EVENT_SEARCH, Bundle().apply {
             putString(Analytics.PARAM_ITEM_ID, translationShortName)