+++ app/src/main/kotlin/me/xizzhu/android/joshua/reading/detail/VerseDetailPresenter.kt
@@ -29,6 +29,7 @@ import me.xizzhu.android.joshua.ui.TranslationInfoComparator
 import me.xizzhu.android.joshua.utils.activities.BaseSettingsPresenter
 import me.xizzhu.android.joshua.utils.supervisedAsync
 import me.xizzhu.android.logger.Log
+import java.lang.StringBuilder
 
 class VerseDetailPresenter(private val readingInteractor: ReadingInteractor)
     : BaseSettingsPresenter<VerseDetailView>(readingInteractor) {
@@ -74,24 +75,61 @@ class VerseDetailPresenter(private val readingInteractor: ReadingInteractor)
                         .sortedWith(translationComparator)
                         .filter { it.shortName != currentTranslation }
                         .map { it.shortName }
-                val verse = readingInteractor.readVerse(currentTranslation, parallelTranslations, verseIndex)
-                val verseTextItems = mutableListOf<VerseTextItem>().apply {
-                    add(VerseTextItem(verseIndex, verse.text,
-                            readingInteractor.readBookNames(verse.text.translationShortName)[verseIndex.bookIndex],
-                            this@VerseDetailPresenter::onVerseClicked,
-                            this@VerseDetailPresenter::onVerseLongClicked))
+                val verses = readingInteractor.readVerses(currentTranslation, parallelTranslations,
+                        verseIndex.bookIndex, verseIndex.chapterIndex)
+
+                // 1. finds the verse
+                var start: VerseIndex? = null
+                for (verse in verses) {
+                    if (verse.text.text.isNotEmpty()) start = verse.verseIndex // we need to consider the empty verses
+                    if (verse.verseIndex.verseIndex >= verseIndex.verseIndex) break
                 }
-                verse.parallel.forEach {
-                    verseTextItems.add(VerseTextItem(verseIndex, it,
-                            readingInteractor.readBookNames(it.translationShortName)[verseIndex.bookIndex],
-                            this@VerseDetailPresenter::onVerseClicked,
-                            this@VerseDetailPresenter::onVerseLongClicked))
+
+                val verseIterator = verses.iterator()
+                var verse: Verse? = null
+                while (verseIterator.hasNext()) {
+                    val v = verseIterator.next()
+                    if (v.verseIndex == start) {
+                        verse = v
+                        break
+                    }
+                }
+                if (verse == null) throw IllegalStateException("Failed to find target verse")
+
+                // 2. builds the parallel
+                val verseTextItems = ArrayList<VerseTextItem>(parallelTranslations.size + 1)
+                val parallel = Array(parallelTranslations.size) { StringBuilder() }
+                val parallelBuilder: (index: Int, Verse.Text) -> Unit = { index, text ->
+                    with(parallel[index]) {
+                        if (isNotEmpty()) append(' ')
+                        append(text.text)
+                    }
+                }
+                verse.parallel.forEachIndexed(parallelBuilder)
+
+                var followingEmptyVerseCount = 0
+                while (verseIterator.hasNext()) {
+                    val v = verseIterator.next()
+                    if (v.text.text.isNotEmpty()) break
+                    v.parallel.forEachIndexed(parallelBuilder)
+                    followingEmptyVerseCount++
+                }
+
+                // 3. constructs VerseTextItems
+                verseTextItems.add(VerseTextItem(verse.verseIndex, followingEmptyVerseCount, verse.text,
+                        readingInteractor.readBookNames(verse.text.translationShortName)[verse.verseIndex.bookIndex],
+                        this@VerseDetailPresenter::onVerseClicked, this@VerseDetailPresenter::onVerseLongClicked))
+
+                parallelTranslations.forEachIndexed { index, translation ->
+                    verseTextItems.add(VerseTextItem(verse.verseIndex, followingEmptyVerseCount,
+                            Verse.Text(translation, parallel[index].toString()),
+                            readingInteractor.readBookNames(translation)[verse.verseIndex.bookIndex],
+                            this@VerseDetailPresenter::onVerseClicked, this@VerseDetailPresenter::onVerseLongClicked))
                 }
 
                 verseDetail = VerseDetail(verseIndex, verseTextItems,
                         bookmarkAsync.await().isValid(), highlightAsync.await().color,
                         noteAsync.await().note)
-
                 view?.onVerseDetailLoaded(verseDetail!!)
             } catch (e: Exception) {
                 Log.e(tag, "Failed to load verse detail", e)