+++ app/src/main/kotlin/me/xizzhu/android/joshua/reading/detail/VerseTextItem.kt
@@ -31,7 +31,8 @@ import me.xizzhu.android.joshua.ui.createBookNameStyleSpan
 import me.xizzhu.android.joshua.ui.recyclerview.BaseItem
 import me.xizzhu.android.joshua.ui.recyclerview.BaseViewHolder
 
-data class VerseTextItem(val verseIndex: VerseIndex, val verseText: Verse.Text, private val bookName: String,
+data class VerseTextItem(val verseIndex: VerseIndex, private val followingEmptyVerseCount: Int,
+                         val verseText: Verse.Text, private val bookName: String,
                          val onClicked: (String) -> Unit, val onLongClicked: (Verse) -> Unit)
     : BaseItem(R.layout.item_verse_text, { inflater, parent -> VerseTextItemViewHolder(inflater, parent) }) {
     companion object {
@@ -54,6 +55,9 @@ data class VerseTextItem(val verseIndex: VerseIndex, val verseText: Verse.Text, 
         SPANNABLE_STRING_BUILDER.append(verseText.translationShortName).append(", ")
                 .append(bookName).append(' ')
                 .append((verseIndex.chapterIndex + 1).toString()).append(':').append((verseIndex.verseIndex + 1).toString())
+        if (followingEmptyVerseCount > 0) {
+            SPANNABLE_STRING_BUILDER.append('-').append((verseIndex.verseIndex + followingEmptyVerseCount + 1).toString())
+        }
         SPANNABLE_STRING_BUILDER.setSpan(BOOK_NAME_STYLE_SPAN, 0, SPANNABLE_STRING_BUILDER.length, Spannable.SPAN_INCLUSIVE_EXCLUSIVE)
         SPANNABLE_STRING_BUILDER.setSpan(BOOK_NAME_SIZE_SPAN, 0, SPANNABLE_STRING_BUILDER.length, Spannable.SPAN_INCLUSIVE_EXCLUSIVE)